streamlit_app/requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
google-cloud-bigquery>=3.23
google-auth>=2.35
db-dtypes>=1.2
pyarrow>=16.1
